Add listen_address to nats service config
Sets up `net:` configuration directive and sets it to `"::"`. Note: the
current default behavior seems that nats is listening on `0.0.0.0`.

Since this module supports only Linux and `net.ipv6.bindv6only=0` is the
default on all of these, `:: `is the expected behaviour as more SRV
records will resolve A and AAAA records these days.