(#57) Don't include class systemd in install/systemd.pp
The `systemd` class is already included by `systemd::unit_file` so it is
unnecessary here.

This also prevents unwanted restarts of the `gnatsd` service that happen
when other classes use defined types from the systemd module, for
example when managing a drop-in file using `systemd::dropin_file`. Such
defined types trigger `Exec[systemctl-daemon-reload]` which propagates
through to the `nats::install::systemd` class which notifies the
`nats::service` class. Not including the `systemd` class here fixes
that.