(#52) Make write_deadline configurable
The write_deadline is how long NATS will wait for a client to read
something before considering it slow and/or dead, when this happen the
client is disconnected and all messages are lost.

Usually this is all fine on 2s but I found on networks approaching 50
000 nodes sometimes you get timeouts and it's worth bumping this up by a
second or 3.  By 300 000 nodes I had to bump it up quite significantly