(#61) Support disabling TLS
While testing with the emulator it was found that on the NATS server
there is a TLS session creation overhead of maybe 0.1s per connection,
this does not seem like a lot but on 100k or more nodes it's a huge
problem.

Apcera and I looked at it and we narrowed this down to ... just how it
is, the golang core libs are this fast and probably it's due to basic
protocol needs.

Go : Go connetions can be a bit faster due to their light weight
protocols but Ruby : Go not.  So this has to be a option.